Rename protocol card list
To avoid confusion with ProtocolList; this specifically displays
protocols from a server.
