+++ src/components/Setup/index.js
@@ -1,6 +1,6 @@
 export { default as PairingCodeInput } from './PairingCodeInput';
 export { default as ProtocolCard } from './ProtocolCard';
-export { default as ProtocolCardList } from './ProtocolCardList';
+export { default as ServerProtocolList } from './ServerProtocolList';
 export { default as ServerCard } from './ServerCard';
 export { default as ServerAddressForm } from './ServerAddressForm';
 export { default as ServerPairingForm } from './ServerPairingForm';
