+++ src/containers/Setup/ServerProtocols.js
@@ -6,7 +6,7 @@ import { Redirect } from 'react-router-dom';
 
 import ApiClient from '../../utils/ApiClient';
 import { actionCreators as protocolActions } from '../../ducks/modules/protocol';
-import { ProtocolCardList, ServerSetup } from '../../components/Setup';
+import { ServerProtocolList, ServerSetup } from '../../components/Setup';
 
 class ServerProtocols extends Component {
   constructor(props) {
@@ -43,7 +43,7 @@ class ServerProtocols extends Component {
     return (
       <ServerSetup server={server}>
         {
-          protocols && <ProtocolCardList protocols={protocols} download={downloadProtocol} />
+          protocols && <ServerProtocolList protocols={protocols} download={downloadProtocol} />
         }
       </ServerSetup>
     );
